requirements.txt
aiogram>=3.0
requests>=2.28
Flask>=2.2
python-dotenv>=0.21  # опционально, если используете .env локально
gitpython>=3.1